+++ carp_core/lib/client/client_manager.dart
@@ -158,10 +158,12 @@ abstract class ClientManager<
   @mustCallSuper
   Future<StudyDeploymentStatus?> getStudyDeploymentStatus(TStudy study) async {
     _checkConfiguration();
-    if (repository.hasStudy(study)) {
-      return await proxy?.getStudyDeploymentStatus(study);
-    }
-    return null;
+    if (!repository.hasStudy(study)) return null;
+
+    final status = study.status;
+    final deploymentStatus = await proxy?.getStudyDeploymentStatus(study);
+    if (study.status != status) repository.updateStudy(study);
+    return deploymentStatus;
   }
 
   /// Verifies whether the device is ready for deployment of the study runtime